requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==2.8.0
djangorestframework==3.13.1
djongo==1.3.6
et-xmlfile==1.1.0
jsonfield==3.1.0
Markdown==3.4.1
MarkupPy==1.14
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.10
pathspec==0.9.0
platformdirs==2.5.2
pymongo==3.12.3
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.2.4
tablib==3.2.1
tomli==2.0.1
xlrd==2.0.1
xlwt==1.3.0
yapf==0.32.0
